Reorganize code for VoteButton and fix imports
