+++ src/components/Vote/VoteButton/SkipButton.js
