+++ src/components/Vote/VoteButton/VoteButton.js
