+++ src/components/Vote/VoteButton/index.js
